requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# AI providers (install the one you need)
openai==1.51.0
google-generativeai==0.8.3

# Optional: for better async support
httpx==0.27.2

